+++ test/core_geometry_test.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 import 'dart:math' as math;
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:iwb_canvas_engine/iwb_canvas_engine.dart';
+import 'package:iwb_canvas_engine/advanced.dart';
 
 void main() {
   test('rotatePoint rotates around center', () {