+++ test/serialization_scene_codec_validation_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:iwb_canvas_engine/iwb_canvas_engine.dart';
+import 'package:iwb_canvas_engine/advanced.dart';
 
 Map<String, dynamic> _minimalSceneJson() {
   return <String, dynamic>{